# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Handles authentication utilities",
        "title": "nrs-authentication",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/auth",
    "paths": {
        "/get-user": {
            "get": {
                "description": "Returns a user by email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User lookup result.",
                        "schema": {
                            "$ref": "#/definitions/dto.ListUsersOutputSwagger"
                        }
                    }
                }
            }
        },
        "/invites/{inviteId}/accept": {
            "get": {
                "description": "Marks an invite as accepted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Accept invite",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invite ID",
                        "name": "inviteId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AcceptInviteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.AcceptInviteResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.AcceptInviteResponse"
                        }
                    }
                }
            }
        },
        "/invites/{inviteId}/attach-role": {
            "post": {
                "description": "Attaches the invited role after the invite is accepted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Attach role from invite",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invite ID",
                        "name": "inviteId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AttachRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.AttachRoleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.AttachRoleResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.AttachRoleResponse"
                        }
                    }
                }
            }
        },
        "/me/attach-role": {
            "post": {
                "description": "attaches a role to a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Attach a role to a user",
                "parameters": [
                    {
                        "description": "Attach role request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AttachRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AttachRoleResponse"
                        }
                    }
                }
            }
        },
        "/me/get-facility-users": {
            "get": {
                "description": "returns a list of facility users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get users for a facility",
                "parameters": [
                    {
                        "type": "string",
                        "description": "facility_code",
                        "name": "facility_code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group",
                        "name": "group",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.UserTypeSwagger"
                            }
                        }
                    }
                }
            }
        },
        "/me/invites": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates an invite, sends the email, and stores invite metadata.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Invite a user by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create invite request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InviteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/me/invites/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all invites that have not been accepted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "List pending invites",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.InviteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AcceptInviteResponse": {
            "type": "object",
            "properties": {
                "invite": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.AttachRoleRequest": {
            "type": "object",
            "required": [
                "groupName",
                "username"
            ],
            "properties": {
                "groupName": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "username": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "dto.AttachRoleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.AttributeTypeSwagger": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                },
                "Value": {
                    "type": "string"
                }
            }
        },
        "dto.CreateInviteRequest": {
            "type": "object",
            "required": [
                "organizationId",
                "organizationName",
                "roleName",
                "targetEmail"
            ],
            "properties": {
                "organizationId": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "organizationName": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "roleName": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "targetEmail": {
                    "type": "string"
                }
            }
        },
        "dto.InviteResponse": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "hostEmail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organizationId": {
                    "type": "string"
                },
                "organizationName": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                },
                "sent": {
                    "type": "boolean"
                },
                "targetEmail": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.ListUsersOutputSwagger": {
            "type": "object",
            "properties": {
                "PaginationToken": {
                    "type": "string"
                },
                "ResultMetadata": {
                    "$ref": "#/definitions/dto.MetadataSwagger"
                },
                "Users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserTypeSwagger"
                    }
                }
            }
        },
        "dto.MFAOptionTypeSwagger": {
            "type": "object",
            "properties": {
                "AttributeName": {
                    "type": "string"
                },
                "DeliveryMedium": {
                    "type": "string"
                }
            }
        },
        "dto.MetadataSwagger": {
            "type": "object"
        },
        "dto.UserTypeSwagger": {
            "type": "object",
            "properties": {
                "Attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AttributeTypeSwagger"
                    }
                },
                "Enabled": {
                    "type": "boolean"
                },
                "MFAOptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MFAOptionTypeSwagger"
                    }
                },
                "UserCreateDate": {
                    "type": "string"
                },
                "UserLastModifiedDate": {
                    "type": "string"
                },
                "UserStatus": {
                    "type": "string"
                },
                "Username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}